Split KnoraUser and User services
