+++ integration/src/test/scala/org/knora/webapi/routing/AuthenticatorSpec.scala
@@ -15,7 +15,6 @@ import org.knora.webapi.messages.v2.routing.authenticationmessages.CredentialsId
 import org.knora.webapi.messages.v2.routing.authenticationmessages.KnoraCredentialsV2.KnoraJWTTokenCredentialsV2
 import org.knora.webapi.messages.v2.routing.authenticationmessages.KnoraCredentialsV2.KnoraPasswordCredentialsV2
 import org.knora.webapi.routing.Authenticator.AUTHENTICATION_INVALIDATION_CACHE_NAME
-import org.knora.webapi.routing.Authenticator.BAD_CRED_NOT_VALID
 import org.knora.webapi.sharedtestdata.SharedTestDataADM
 import org.knora.webapi.slice.admin.domain.model.Email
 import org.knora.webapi.slice.admin.domain.model.User
@@ -35,19 +34,6 @@ class AuthenticatorSpec extends CoreSpec with ImplicitSender with PrivateMethodT
   private def testUserAdmFromIri(iri: String) = User(iri, "", "", "", "", false, "")
 
   "During Authentication" when {
-    "called, the 'getUserADMByEmail' method " should {
-      "succeed with the correct 'email' " in {
-        val user =
-          UnsafeZioRun.runOrThrow(Authenticator.getUserByEmail(Email.unsafeFrom(AuthenticatorSpec.rootUserEmail)))
-        assert(user == AuthenticatorSpec.rootUser)
-      }
-
-      "fail with the wrong 'email' " in {
-        val actual = UnsafeZioRun.run(Authenticator.getUserByEmail(Email.unsafeFrom("wronguser@example.com")))
-        assertFailsWithA[BadCredentialsException](actual, BAD_CRED_NOT_VALID)
-      }
-    }
-
     "called, the 'authenticateCredentialsV2' method" should {
       "succeed with correct email/password" in {
         val credId               = CredentialsIdentifier.EmailIdentifier(Email.unsafeFrom(AuthenticatorSpec.rootUserEmail))
