+++ webapi/src/main/scala/org/knora/webapi/core/LayersLive.scala
@@ -5,9 +5,6 @@
 
 package org.knora.webapi.core
 
-import zio.ULayer
-import zio.ZLayer
-
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.config.AppConfig.AppConfigurations
 import org.knora.webapi.config.InstrumentationServerConfig
@@ -65,6 +62,8 @@ import org.knora.webapi.store.iiif.impl.SipiServiceLive
 import org.knora.webapi.store.triplestore.api.TriplestoreService
 import org.knora.webapi.store.triplestore.impl.TriplestoreServiceLive
 import org.knora.webapi.store.triplestore.upgrade.RepositoryUpdater
+import zio.ULayer
+import zio.ZLayer
 
 object LayersLive {
 
@@ -72,18 +71,19 @@ object LayersLive {
    * The `Environment` that we require to exist at startup.
    */
   type DspEnvironmentLive =
-    ActorSystem & AdminApiEndpoints & ApiRoutes & ApiV2Endpoints & AppConfigurations & AppRouter & Authenticator &
-      AuthorizationRestService & CacheService & CacheServiceRequestMessageHandler & CardinalityHandler &
-      CardinalityService & ConstructResponseUtilV2 & ConstructTransformer & GravsearchTypeInspectionRunner &
-      GroupsResponderADM & HttpServer & IIIFRequestMessageHandler & InferenceOptimizationService &
-      InstrumentationServerConfig & IriConverter & IriService & JwtService & KnoraProjectRepo & ListsResponder &
-      ListsResponderV2 & MessageRelay & OntologyCache & OntologyHelpers & OntologyInferencer & OntologyRepo &
-      OntologyResponderV2 & PasswordService & PermissionsResponderADM & PermissionsRestService & PermissionUtilADM &
-      PredicateObjectMapper & ProjectADMRestService & ProjectADMService & ProjectExportService &
-      ProjectExportStorageService & ProjectImportService & ProjectsResponderADM & QueryTraverser & RepositoryUpdater &
-      ResourcesResponderV2 & ResourceUtilV2 & ResourceUtilV2 & RestCardinalityService & RestResourceInfoService &
-      SearchApiRoutes & SearchResponderV2 & AssetPermissionsResponder & SipiService & StandoffResponderV2 & StandoffTagUtilV2 &
-      State & StoreRestService & StringFormatter & TriplestoreService & UsersResponder & ValuesResponderV2 & UsersRestService & UserService
+    ActorSystem & AdminApiEndpoints & ApiRoutes & ApiV2Endpoints & AppConfigurations & AppRouter &
+      AssetPermissionsResponder & Authenticator & AuthorizationRestService & CacheService &
+      CacheServiceRequestMessageHandler & CardinalityHandler & CardinalityService & ConstructResponseUtilV2 &
+      ConstructTransformer & GravsearchTypeInspectionRunner & GroupsResponderADM & HttpServer &
+      IIIFRequestMessageHandler & InferenceOptimizationService & InstrumentationServerConfig & IriConverter &
+      IriService & JwtService & KnoraProjectRepo & KnoraUserService & ListsResponder & ListsResponderV2 &
+      MessageRelay & OntologyCache & OntologyHelpers & OntologyInferencer & OntologyRepo & OntologyResponderV2 &
+      PasswordService & PermissionsResponderADM & PermissionsRestService & PermissionUtilADM & PredicateObjectMapper &
+      ProjectADMRestService & ProjectADMService & ProjectExportService & ProjectExportStorageService &
+      ProjectImportService & ProjectsResponderADM & QueryTraverser & RepositoryUpdater & ResourcesResponderV2 &
+      ResourceUtilV2 & ResourceUtilV2 & RestCardinalityService & RestResourceInfoService & SearchApiRoutes &
+      SearchResponderV2 & SipiService & StandoffResponderV2 & StandoffTagUtilV2 & State & StoreRestService &
+      StringFormatter & TriplestoreService & UserService & UsersResponder & UsersRestService & ValuesResponderV2
 
   /**
    * All effect layers needed to provide the `Environment`
@@ -124,9 +124,10 @@ object LayersLive {
       JwtServiceLive.layer,
       KnoraProjectRepoLive.layer,
       KnoraResponseRenderer.layer,
-      KnoraUserRepoLive.layer,
-      KnoraUserToUserConverter.layer,
       KnoraUserGroupRepoLive.layer,
+      KnoraUserRepoLive.layer,
+      KnoraUserService.layer,
+      KnoraUserToUserConverter.layer,
       ListRestService.layer,
       ListsEndpoints.layer,
       ListsEndpointsHandlers.layer,
