+++ webapi/src/main/scala/org/knora/webapi/messages/util/UserUtilADM.scala
@@ -6,7 +6,6 @@
 package org.knora.webapi.messages.util
 
 import zio.*
-
 import dsp.errors.ForbiddenException
 import dsp.errors.NotFoundException
 import org.knora.webapi.IRI
