+++ webapi/src/main/scala/org/knora/webapi/routing/v2/ResourcesRouteV2.scala
@@ -12,7 +12,6 @@ import org.apache.pekko.http.scaladsl.server.Route
 import zio.*
 
 import java.time.Instant
-
 import dsp.errors.BadRequestException
 import dsp.valueobjects.Iri
 import org.knora.webapi.*
